refactor: extract CargoField into separate file
