+++ src/pages-components/Dashboard/ShipmentsPage/NewShipmentPage/NewShipmentForm.tsx
@@ -1,9 +1,10 @@
 import { FormProvider, useForm, useFormContext } from 'react-hook-form';
 
 import { FormSingleSelect, FormTextInput } from '@/lib/components/form';
-import { FormTextarea } from '@/lib/components/form/FormTextarea';
 import { postalCodes } from '@/lib/mocks/postalCodes';
-import { Box, Button, Divider, FlexLayout, Icon, Text, TextButton } from '@/ui';
+import { Box, Divider, FlexLayout, Text, TextButton } from '@/ui';
+
+import { CargoField } from './CargoField';
 
 const defaultValues: {
   orderNo: string;
@@ -109,8 +110,8 @@ export const NewShipmentForm = () => {
             <Divider />
           </Box>
           <FlexLayout className="flex-col gap-4">
-            {cargo.map((_, index: number) => (
-              <CargoField index={index} key={index} />
+            {cargo.map((_, index: number, arr: any[]) => (
+              <CargoField cargoLength={arr.length} index={index} key={index} />
             ))}
             <TextButton
               iconLeft="PlusIcon"
@@ -156,122 +157,6 @@ export const NewShipmentForm = () => {
         <pre className="text-xs whitespace-pre-wrap">{JSON.stringify(formValues, null, 2)}</pre>
       </Box>
     </FormProvider>
-  );
-};
-
-const CargoField = ({ index }: { index: number }) => {
-  const { watch, setValue, getValues } = useFormContext();
-  const cargoType = watch(`cargo.${index}.metadata.type`);
-  const isStandardCargo = cargoType === 'standard';
-
-  const setCargoType = (type: 'standard' | 'nonstandard') => {
-    setValue(`cargo.${index}.metadata.type`, type);
-  };
-
-  const removeCargo = () => {
-    const currentCargo = getValues('cargo');
-    const updatedCargo = currentCargo.filter((_, i) => i !== index);
-    setValue('cargo', updatedCargo);
-  };
-
-  const currentCargo = getValues('cargo');
-  const cargoLength = currentCargo.length;
-
-  return (
-    <FlexLayout as="fieldset" className="flex-col gap-4 bg-black-alpha-10 dark:bg-white-alpha-10 p-4 rounded-s">
-      <FlexLayout className="justify-between items-center">
-        <Text color="text-color-3" variant="text-s-medium">
-          TERET {index + 1}
-        </Text>
-        {cargoLength > 1 && <Icon icon="XMarkIcon" onClick={removeCargo} />}
-      </FlexLayout>
-      <FlexLayout className="flex-col gap-4">
-        <FlexLayout className="gap-2">
-          <Box className="flex-1">
-            <Button
-              isFullWidth
-              text="Standardni teret"
-              type="button"
-              variant={isStandardCargo ? 'primary' : 'secondary'}
-              onClick={() => setCargoType('standard')}
-            />
-          </Box>
-          <Box className="flex-1">
-            <Button
-              isFullWidth
-              text="Nestandardni teret"
-              type="button"
-              variant={!isStandardCargo ? 'primary' : 'secondary'}
-              onClick={() => setCargoType('nonstandard')}
-            />
-          </Box>
-        </FlexLayout>
-        {isStandardCargo ? (
-          <FlexLayout className="gap-4">
-            <Box className="flex-1">
-              <FormSingleSelect
-                isClearable
-                isSearchable
-                label="Vrsta palete"
-                name={`cargo.${index}.metadata.palleteType`}
-                options={[
-                  { value: '80x60', label: 'Mala Paleta (80x60)' },
-                  { value: '120x80', label: 'Euro Paleta (120x80)' },
-                  { value: '120x100', label: 'Brodska Paleta (120x100)' },
-                  { value: '100x100', label: 'Industrijska Paleta (100x100)' },
-                  { value: '120x120', label: 'Jumbo Paleta (120x120)' },
-                ]}
-                placeholder="Odaberi vrstu palete..."
-              />
-            </Box>
-            <Box className="flex-1">
-              <FormTextInput
-                label="Broj paleta"
-                min="1"
-                name={`cargo.${index}.metadata.palleteAmount`}
-                placeholder="Unesi broj paleta"
-                type="number"
-              />
-            </Box>
-          </FlexLayout>
-        ) : (
-          <FlexLayout className="gap-4">
-            <Box className="flex-1">
-              <FormTextInput
-                label="Duljina (m)"
-                min="0"
-                name={`cargo.${index}.metadata.length`}
-                placeholder="XXX"
-                step="0.01"
-                type="number"
-              />
-            </Box>
-            <Box className="flex-1">
-              <FormTextInput
-                label="Širina (m)"
-                min="0"
-                name={`cargo.${index}.metadata.width`}
-                placeholder="XXX"
-                step="0.01"
-                type="number"
-              />
-            </Box>
-            <Box className="flex-1">
-              <FormTextInput
-                label="Visina (m)"
-                min="0"
-                name={`cargo.${index}.metadata.height`}
-                placeholder="XXX"
-                step="0.01"
-                type="number"
-              />
-            </Box>
-          </FlexLayout>
-        )}
-        <FormTextInput label="Težina (kg)" name={`cargo.${index}.weight`} />
-        <FormTextarea label="Opis tereta" name={`cargo.${index}.description`} />
-      </FlexLayout>
-    </FlexLayout>
   );
 };
